requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.32.0
PyYAML>=6.0.2
Jinja2>=3.1.4
GitPython>=3.1.43
